% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{imed_enrich_measurements}
\alias{imed_enrich_measurements}
\title{Join measurement records to their scientific comparability metadata}
\usage{
imed_enrich_measurements(x)
}
\arguments{
\item{x}{An imed_db.}
}
\description{
Join measurement records to their scientific comparability metadata
}
