% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_reporting.R
\name{imed_evidence_profile}
\alias{imed_evidence_profile}
\title{Evidence-quality/reporting profile}
\usage{
imed_evidence_profile(x)
}
\arguments{
\item{x}{An `imed_db` object, as returned by `imed_data()`.}
}
\description{
Evidence-quality/reporting profile
}
