% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{imed_flag_holds}
\alias{imed_flag_holds}
\title{Flag measurement-level scientific HOLDs}
\usage{
imed_flag_holds(x)
}
\arguments{
\item{x}{An `imed_db`, or a measurement-level data frame that includes comparability
metadata.}
}
\description{
Flag measurement-level scientific HOLDs
}
