% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_validation.R
\name{imed_forest}
\alias{imed_forest}
\title{Forest plot for an imedR meta-analysis}
\usage{
imed_forest(fit, ...)
}
\arguments{
\item{fit}{A fitted `metafor` `rma` object, for example from `imed_meta_mean()`.}

\item{...}{Further arguments passed to `metafor::forest()`.}
}
\description{
Forest plot for an imedR meta-analysis
}
