% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_reporting.R
\name{imed_manuscript_table}
\alias{imed_manuscript_table}
\title{Study-characteristics table for manuscripts}
\usage{
imed_manuscript_table(x, study_id = NULL)
}
\arguments{
\item{x}{An `imed_db` object, as returned by `imed_data()`.}

\item{study_id}{Optional character vector of `Study_ID` values used to restrict the records.}
}
\description{
Study-characteristics table for manuscripts
}
