% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{imed_plot_composition}
\alias{imed_plot_composition}
\title{Plot polymer or particle-characteristic composition}
\usage{
imed_plot_composition(
  x,
  what = c("polymers", "particles"),
  study_id = NULL,
  characteristic_type = NULL
)
}
\arguments{
\item{x}{An `imed_db` object, as returned by `imed_data()`.}

\item{what}{polymers or particles.}

\item{study_id}{Optional character vector of `Study_ID` values used to restrict the records.}

\item{characteristic_type}{Optional character vector of `Characteristic_Type` values; used only when
`what = "particles"`.}
}
\description{
Plot polymer or particle-characteristic composition
}
