% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{imed_plot_methods}
\alias{imed_plot_methods}
\title{Plot method/QAQC reporting profile}
\usage{
imed_plot_methods(x)
}
\arguments{
\item{x}{An `imed_db` object, as returned by `imed_data()`.}
}
\description{
Plot method/QAQC reporting profile
}
