% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_reporting.R
\name{imed_results_table}
\alias{imed_results_table}
\title{Tidy a statistical result object for manuscript reporting}
\usage{
imed_results_table(x)
}
\arguments{
\item{x}{A result object: an `imed_stats` result or a fitted `metafor` `rma` object.}
}
\description{
Tidy a statistical result object for manuscript reporting
}
