% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{imed_tables}
\alias{imed_tables}
\title{List loaded IMED tables}
\usage{
imed_tables(x = imed_data(validate = FALSE))
}
\arguments{
\item{x}{An `imed_db` object. Defaults to the bundled snapshot loaded without
validation.}
}
\description{
List loaded IMED tables
}
