% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{imed_validate}
\alias{imed_validate}
\title{Validate IMED core scientific and relational integrity}
\usage{
imed_validate(x, strict = FALSE)
}
\arguments{
\item{x}{An imed_db.}

\item{strict}{Stop on ERROR-level failures.}
}
\description{
Validate IMED core scientific and relational integrity
}
